Add the export impot method to the app modules
