+++ index.html
@@ -50,6 +50,6 @@
       <p class="copyright-paragraph">Copyright - Awesome Books 2022</p>
     </footer>
     <script src="https://cdn.jsdelivr.net/npm/luxon@2/build/global/luxon.min.js"></script>
-    <script src="index.js"></script>
+    <script src="./modules/index.js" type="module"></script>
   </body>
 </html>
